fix(verify-series): ensure correct ids verified
When the SeriesFile in 2.0 was changed to include typed information
about the series this tool was not updated. This meant it was trying to
verify the typed series id (which use bits > 32-bit) were monotonically
increasing.

The invariant is not true because it depends on the type of the series
as to what bits are set. This commit ensures that the underlying series
id (with type information removed) is verfied.