feat(verify-seriesfile): add new options to tool
This commit adds two new options to the `influxd inspect
verify-seriesfile` command.

The first is the ability to surpress panic recovery: `--recover=false`.
This is useful when you want to understand where in the codebase the
panic happened. The second is the ability to continue verification after
encountering _some_ errors (not all types of error should be continued
from). This flag is `--continue-on-error`; it defaults to `false`.